# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Commutrum — JSE factor research engine API
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /universe:
    get:
      operationId: getUniverse
      tags: [universe]
      summary: Get investable universe for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Universe snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniverseListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /universe/{ticker}/screen-status:
    get:
      operationId: getTickerScreenStatus
      tags: [universe]
      summary: Get screen status for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Screen status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Module 01 — Data ─────────────────────────────────────────────────────
...
components:
  parameters:
    ticker:
      name: ticker
      in: path
      required: true
      schema:
        type: string
...
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 100
        maximum: 500
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Meta:
      type: object
      properties:
        as_of:
          type: string
          format: date
        generated_at:
          type: string
          format: date-time
      required: [generated_at]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    UniverseEntry:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: string
        marketCapZar:
          type: number
        adtv20d:
          type: number
        isActive:
          type: boolean
      required: [ticker, isActive]
    UniverseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UniverseEntry"
        meta:
          $ref: "#/components/schemas/Meta"
        nextCursor:
          type: string
      required: [data, meta]
    ScreenStatusResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticker:
              type: string
            asOf:
              type: string
              format: date
            passes:
              type: boolean
            filters:
              type: object
              additionalProperties:
                type: boolean
          required: [ticker, passes]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...